# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Credit Repair Pro API",
    "description": "Complete API for Credit Repair SaaS — client management, credit tracking, dispute generation, payments, compliance (CROA/FCRA/GLBA), and AI-powered analysis.",
    "version": "1.0.0",
    "contact": {
      "name": "TriExpert Services",
      "url": "https://triexpertservice.com"
    }
  },
  "servers": [
    { "url": "https://triexpertservice.com/api", "description": "Production" },
    { "url": "http://localhost:5000/api", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Local JWT or Auth0 token"
      }
    },
...
  },
  "paths": {
...
    "/subscriptions/admin/transactions": {
      "get": { "tags": ["Subscriptions"], "summary": "Get all transactions", "operationId": "getTransactions", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }, { "name": "offset", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Transaction list" } } }
    },
...
    "/plaid/create-link-token": { "post": { "tags": ["Plaid"], "summary": "Create Plaid Link token", "operationId": "createLinkToken", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Link token" } } } },
    "/plaid/exchange-token": { "post": { "tags": ["Plaid"], "summary": "Exchange public token for access", "operationId": "exchangeToken", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["publicToken"], "properties": { "publicToken": { "type": "string" }, "metadata": { "type": "object" } } } } } }, "responses": { "200": { "description": "Token exchanged" } } } },
    "/plaid/accounts": { "get": { "tags": ["Plaid"], "summary": "Get linked bank accounts", "operationId": "getPlaidAccounts", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Account list" } } } },
    "/plaid/verification-status": { "get": { "tags": ["Plaid"], "summary": "Get identity verification status", "operationId": "getVerificationStatus", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Verification status" } } } },
    "/plaid/refresh-identity": { "post": { "tags": ["Plaid"], "summary": "Refresh identity verification", "operationId": "refreshIdentity", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Identity refreshed" } } } },
    "/plaid/get-transactions": { "post": { "tags": ["Plaid"], "summary": "Get recent transactions", "operationId": "getTransactionsPlaid", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "days": { "type": "integer", "default": 30 } } } } } }, "responses": { "200": { "description": "Transactions" } } } },
    "/plaid/accounts/{itemId}": { "delete": { "tags": ["Plaid"], "summary": "Remove linked bank account", "operationId": "removePlaidAccount", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "itemId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Account removed" } } } },
...
  }
}